Update Balloons.kt
This one was awful. There was one stupid bug that I was working on for like an hour and a half and I ended up just fixing it by just offsetting the user's tap coordinates to the right place. Also I added the ability to tap balloons to pop them.